State manager, Currency page
A state manager has been added that retrieves data from the API. Additionally, an API key has been added to the .env file, which will be pushed to Git along with the project so that it can be checked without unnecessary clarification of the key. The state manager has been enhanced to store the fetched data, and functionality has been added to prevent requests from being sent more often than once every 5 minutes. If requests are more frequent, the displayed data will be outdated. A UI for the currency list has been created