updated screens and saving logic
I added a screen with the current exchange rate, and included a chart with mock data since to get this from an API you have to pay money, as all the APIs for this test task require payment to get the history of the current rate. I also added functionality to save a favorite exchange rate to a specific one, and I might still change the logic here to get the rate. When the general list of rates is updated, the current favorite rates will also be updated to keep the data current