Styles Optimization and Data Update
Since the value is used multiple times, it makes sense to move Dimensions.get('screen').width into a separate constant to reduce computations.

According to the documentation, the data is updated once per hour, so there is no point in refreshing it every 10 minutes. Now, it will update every 30 minutes to conserve the API limit